// src/models/signup.ts
export class Signup {
    name: string;
    lastName: string;
    displayName: string;
    birthdate: Date;
    email: string;
    password: string;
    passwordConfirmation: string;
    phone: string;
    cellPhone: string;
    address1: string;
    address2: string;
    zipCode: string;
    neighborhood: string;
    stateId: number;
    cityId: number;
    serviceId: number;
    facebookId: String;
    facebookAcessToken: String;
}